Add maintainer guidelines
This commit adds guidelines for maintainers of fedora-happiness-packets.
The guidelines document tasks for us to keep things running smoothly and
reviewing new changes.

This is not a complete document, but I figured we can start with this
and add new things in as they come up. But this gives us a base to start
with for now. 🌈