Improve client secrets generation script and docs
This commit simplifies the `generate_client_secrets.sh` script by adding
comments, breaking it up across multiple lines, and emitting a message
when secrets already exist. I also remove an unnecessary step from the
development environment instructions to set the executable bit on the
script, since it is already set in the repository.