Add CommOps contact info; remove duplicated testing instructions
This commit adds info on how to contact the CommOps team for working on
this project. It links to our existing docs site where we offer points
of contact.

Additionally, the testing instructions are duplicated in the README and
our project docs. It's best to have them written in one place only. So I
removed them from our README.