dependencies: Start using pipenv for dependency management
This commit adds a Pipfile with some of the dependencies Fedora
Happiness Packets uses. See more about pipenv here:

    https://fedoramagazine.org/install-pipenv-fedora/

I am making this change to help us migrate to a single set of
dependencies. Currently, there are multiple sets of dependencies for
development, production, and the CI environment. The issue with this is
we could have different versions of software at the same time. For
example, a passing build in a CI environment might not work in
production or vice versa.

Ultimately, we should strive to use a single set of dependencies across
the project. pipenv adds a nice front-end to managing dependencies and
helps us manage them at the top-level of the project. For now, this is
just a start but we should work on migrating the requirements here over
time.